+++ src/Products/PloneMeeting/events.py
@@ -549,6 +549,7 @@ def item_added_or_initialized(item):
        To manage every cases, we need to do this in both Initialized and Added event
        because some are triggered in some cases and not others...
        Especially for plone.restapi that calls Initialized then do the validation.'''
+    # make sure workflow mapping is applied, plone.restapi needs it...
     user = api.user.get_current()
     item.manage_addLocalRoles(user.getId(), ('MeetingMember',))
     # Add a place to store adviceIndex